# docs/swagger.yaml
basePath: /api/v1
definitions:
  models.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        type: integer
      error:
        type: string
      success:
        type: boolean
    type: object
...
  models.Message:
    properties:
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_read:
        type: boolean
      message_type:
        description: text, image, gif
        type: string
      sender_id:
        type: string
    type: object
  models.MessageRequest:
    properties:
      content:
        maxLength: 500
        type: string
      message_type:
        enum:
        - text
        - image
        - gif
        type: string
    required:
    - content
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@matching-api.example.com
    name: API Support
    url: https://matching-api.example.com/support
  description: A complete Go-based REST API for a dating application with real-time
    chat, sophisticated matching algorithms, and comprehensive user management.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://matching-api.example.com/terms
  title: Matching API - Matching App Backend
  version: "1.0"
paths:
...
  /chats/{chatID}/messages:
    get:
      consumes:
      - application/json
      description: Retrieve messages for a specific chat conversation with pagination
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        type: string
      - default: 50
        description: Number of messages per page (max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Messages retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    messages:
                      items:
                        $ref: '#/definitions/models.Message'
                      type: array
                    offset:
                      type: integer
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad request - invalid chat ID
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - no access to chat
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Chat not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get chat messages
      tags:
      - Chat
    post:
      consumes:
      - application/json
      description: Send a message in a chat conversation
      parameters:
      - description: Chat ID
        in: path
        name: chatID
        required: true
        type: string
      - description: Message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.MessageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Message sent successfully
          schema:
            allOf:
            - $ref: '#/definitions/models.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Message'
              type: object
        "400":
          description: Bad request - validation failed
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized - invalid token
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden - no access to chat
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send message
      tags:
      - Chat
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"